+++ js/templates/mistakes.js
@@ -1,0 +1,6 @@
+export default () => {
+  return `<div class="main-mistakes">
+    <img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">
+    <img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">
+  </div>`;
+};
