+++ js/templates/welcome.js
@@ -1,5 +1,5 @@
-import {getElementFromTemplate} from '../../utils';
-import showScreen from '../../show-screen';
+import {getElementFromTemplate} from '../utils';
+import showScreen from '../show-screen';
 import getArtistScreen from './artist';
 
 const template = `<!-- Приветствие -->
@@ -20,7 +20,23 @@ export default () => {
   const screen = welcomeScreen.cloneNode(true);
 
   const playBtn = screen.querySelector(`.main-play`);
-  playBtn.addEventListener(`click`, () => showScreen(getArtistScreen()));
+
+
+  playBtn.addEventListener(`click`, () => showScreen(getArtistScreen({
+    answers: [{
+      number: 1,
+      image: `http://placehold.it/134x134`,
+      artist: `Пелагея`
+    }, {
+      number: 2,
+      image: `http://placehold.it/134x134`,
+      artist: `Краснознаменная дивизия имени моей бабушки`
+    }, {
+      number: 3,
+      image: `http://placehold.it/134x134`,
+      artist: `Lorde`
+    }]
+  })));
 
   return screen;
 };
